+++ ghost/core/package.json
@@ -93,7 +93,6 @@
     "@tryghost/email-suppression-list": "0.0.0",
     "@tryghost/errors": "1.3.5",
     "@tryghost/express-dynamic-redirects": "0.0.0",
-    "@tryghost/external-media-inliner": "0.0.0",
     "@tryghost/ghost": "0.0.0",
     "@tryghost/helpers": "1.1.90",
     "@tryghost/html-to-plaintext": "0.0.0",
@@ -183,6 +182,7 @@
     "express-query-boolean": "2.0.0",
     "express-queue": "0.0.13",
     "express-session": "1.18.1",
+    "file-type": "16.5.4",
     "fs-extra": "11.3.0",
     "ghost-storage-base": "1.0.0",
     "glob": "8.1.0",
@@ -213,6 +213,7 @@
     "metascraper-publisher": "5.45.10",
     "metascraper-title": "5.45.10",
     "metascraper-url": "5.45.10",
+    "mime-types": "2.1.35",
     "moment": "2.24.0",
     "moment-timezone": "0.5.45",
     "multer": "1.4.4",
