+++ ghost/core/test/unit/server/services/media-inliner/test/fixtures/fixture.exe
